Revert "Add internal optimize plugin (#883)"
This reverts commit 8186aab07dc6333d6e47775b15e80dc0073ea845.
