--- plugins/plugin-optimize/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "@snowpack/plugin-optimize",
-  "version": "0.0.1",
-  "main": "plugin.js",
-  "dependencies": {
-    "es-module-lexer": "^0.3.24",
-    "esbuild": "^0.6.28",
-    "glob": "^7.1.6",
-    "kleur": "^4.1.1",
-    "p-queue": "^6.6.1"
-  }
-}
