+++ snowpack/package.json
@@ -67,8 +67,8 @@
     "css-modules-loader-core": "^1.1.0",
     "deepmerge": "^4.2.2",
     "detect-port": "^1.3.0",
-    "es-module-lexer": "^0.3.24",
-    "esbuild": "^0.6.28",
+    "es-module-lexer": "^0.3.17",
+    "esbuild": "^0.6.11",
     "etag": "^1.8.1",
     "execa": "^4.0.3",
     "find-cache-dir": "^3.3.1",
@@ -78,12 +78,12 @@
     "http-proxy": "^1.18.1",
     "is-builtin-module": "^3.0.0",
     "jsonschema": "^1.2.5",
-    "kleur": "^4.1.1",
+    "kleur": "^4.1.0",
     "mime-types": "^2.1.26",
     "mkdirp": "^1.0.3",
     "npm-run-path": "^4.0.1",
     "open": "^7.0.4",
-    "p-queue": "^6.6.1",
+    "p-queue": "^6.2.1",
     "resolve-from": "^5.0.0",
     "rimraf": "^3.0.0",
     "rollup": "^2.23.0",
